backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.2.1
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
distro==1.8.0
dnspython==2.4.2
fastapi==0.104.1
fonttools==4.44.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.5
itsdangerous==2.1.2
kiwisolver==1.4.5
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
openai==1.3.5
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.4
uvicorn==0.24.0.post1
